족보공유, 자유게시판, 고민상담, QNA 댓글 달기 및 삭제 각 고유아이디마다 댓글달기 삭제하기 성공.. 수정은 내일 하기로...

+++ project/WebContent/jinju/Counsel/counselSelect.jsp
@@ -3,7 +3,7 @@
 <%@ page import = "miniproject_jj.*" %>
 <%@page import="java.util.*"%>
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
-
+<%@ include file="../../memberCheck.jsp"%>
 <!DOCTYPE html>
 <html>
 <head>
@@ -91,7 +91,7 @@ function chk(){
 					<textarea cols="100%" name = "content"></textarea>
 					<input type="hidden" name="bnum" value=<%=bnum%>> 
 					<input type="hidden" name="pageNum" value=<%=pageNum %>>
-				
+					<input type="hidden" name="writer" value="<%=writer %>">
 					<div class="btnArea">
 						<span class="btnR">
 							<input type="submit" value="댓글달기" class="btnTy3">
@@ -104,24 +104,33 @@ function chk(){
 	<div class = "contents">
 	<table class="tbTy1 detaLeft">
 		<%
-			CounselBoardDAO cbdd = CounselBoardDAO.getInstance();
 			/* replyResult = cbd.inputRipple(bnum); */
-			replyList = cbdd.listRippleSelect(bnum);
+			replyList = cbd.listRippleSelect(bnum);
 			for(int i = 0 ; i < replyList.size(); i++){
 				int width = replyList.get(i).getRe_level() * 10;
 				
-				out.println("<tr><td> 작성자 : *****" + "</td>");
+				out.println("<tr><td> 작성자 :" + replyList.get(i).getWriter() + "</td>");
 				out.println("<td> 날짜 : " + replyList.get(i).getR_date() + "</td>");
-				out.println("<td> <input type = 'button' value = '수정'"+
-						"onclick = location.href='homeMainPage.jsp?pgm=../jinju/CounselReply/counselUpdateReply.jsp?bnum=" +
-						bnum + "&pageNum=" + pageNum + "&re_step=" + replyList.get(i).getRe_step() + 
-						"'></td></tr>");
-				//두번째 줄
+				
+				if(writer.equals(replyList.get(i).getWriter())){
+					out.println("<td> <input type = 'button' value = '수정'"+
+							"onclick = location.href='homeMainPage.jsp?pgm=../jinju/CounselReply/counselUpdateReply.jsp?bnum=" +
+							bnum + "&pageNum=" + pageNum + "&re_step=" + replyList.get(i).getRe_step() + 
+							"'></td></tr>");
+				}else{
+					//out.println("<td>"+writer+"는 get출력을 하는" + replyList.get(i).getWriter() +"와 같냐?</td></tr>");
+					out.println("<td>"+writer+","+replyList.get(i).getWriter()+"</td></tr>");
+				}
 				out.println("<tr><td colspan = '2'>" + replyList.get(i).getContent() + "</td>");
-				out.println("<td> <input type = 'button' value = '삭제'"+
-						"onclick = location.href='homeMainPage.jsp?pgm=../jinju/CounselReply/counselDeleteReplyPro.jsp?bnum=" +
-						bnum + "&pageNum=" + pageNum + "&re_step=" + replyList.get(i).getRe_step() + 
-						"'></td></tr>");
+				if(writer.equals(replyList.get(i).getWriter())){
+					out.println("<td> <input type = 'button' value = '삭제'"+
+							"onclick = location.href='homeMainPage.jsp?pgm=../jinju/CounselReply/counselDeleteReplyPro.jsp?bnum=" +
+							bnum + "&pageNum=" + pageNum + "&re_step=" + replyList.get(i).getRe_step() + 
+							"'></td></tr>");
+				}else{
+					out.println("<td></td></tr>");
+				}
+				
 				out.println("</tr>");
 				/* if (replyList.get(i).getRe_level() > 0) {
 					out.println("<img src='images/level.gif' width="
